feat: adding endpoints for search and delete in Review entity, allowing access reviews by current user and userId and delete reviews by current user; fix: fixing ENUM of the class EndpointsPermissions to increment a AUTHENTICATED option

+++ backend/spring-boot/src/main/java/com/project/cinescope/application/review/impl/ReviewServiceImpl.java
@@ -1,5 +1,6 @@
 package com.project.cinescope.application.review.impl;
 
+import com.project.cinescope.application.auth.service.AuthenticatedUserService;
 import com.project.cinescope.application.movie.Movie;
 import com.project.cinescope.application.movie.MovieRepository;
 import com.project.cinescope.application.review.Review;
@@ -9,6 +10,7 @@ import com.project.cinescope.application.review.request.ReviewRequestDto;
 import com.project.cinescope.application.review.response.ReviewResponseDto;
 import com.project.cinescope.application.user.User;
 import com.project.cinescope.application.user.UserRepository;
+import com.project.cinescope.core.exception.exceptions.ForbiddenOperationException;
 import com.project.cinescope.core.exception.exceptions.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
@@ -19,11 +21,13 @@ public class ReviewServiceImpl implements ReviewService {
     private final ReviewRepository reviewRepository;
     private final UserRepository userRepository;
     private final MovieRepository movieRepository;
+    private final AuthenticatedUserService authenticatedUserService;
 
-    public ReviewServiceImpl(ReviewRepository reviewRepository, UserRepository userRepository, MovieRepository movieRepository) {
+    public ReviewServiceImpl(ReviewRepository reviewRepository, UserRepository userRepository, MovieRepository movieRepository, AuthenticatedUserService authenticatedUserService) {
         this.reviewRepository = reviewRepository;
         this.userRepository = userRepository;
         this.movieRepository = movieRepository;
+        this.authenticatedUserService = authenticatedUserService;
     }
 
     public List<ReviewResponseDto> getAll() {
@@ -39,9 +43,26 @@ public class ReviewServiceImpl implements ReviewService {
                 .orElseThrow(() -> new ResourceNotFoundException("Review not found with id " + id));
     }
 
+    public List<ReviewResponseDto> findReviewsByCurrentUser() {
+        User user = authenticatedUserService.getCurrentUser();
+        List<Review> reviewList = reviewRepository.findByUser(user);
+        return reviewList.stream()
+                .map(ReviewResponseDto::toReviewDto)
+                .toList();
+    }
+
+    public List<ReviewResponseDto> findReviewsByUserId(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userId));
+        List<Review> reviewList = reviewRepository.findByUser(user);
+        return reviewList.stream()
+                .map(ReviewResponseDto::toReviewDto)
+                .toList();
+    }
+
     public ReviewResponseDto post(ReviewRequestDto requestDto) {
-        User user = userRepository.findById(requestDto.userId())
-                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + requestDto.userId()));
+        User user = authenticatedUserService.getCurrentUser();
+
         Movie movie = movieRepository.findById(requestDto.movieId())
                 .orElseThrow(() -> new ResourceNotFoundException("Movie not found with id: " + requestDto.movieId()));
 
@@ -56,6 +77,18 @@ public class ReviewServiceImpl implements ReviewService {
         Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Review not found with id: " + id));
 
+        reviewRepository.delete(review);
+    }
+
+    public void deleteReviewByCurrentUser(Long id) {
+        Review review = reviewRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Review not found with id: " + id));
+
+        User user = authenticatedUserService.getCurrentUser();
+
+        if (user != review.getUser()) {
+            throw new ForbiddenOperationException("You don't have permission to delete another user's reviews.");
+        }
         reviewRepository.delete(review);
     }
 }
